simple-backend/requirements_dev.txt
setuptools~=57.4.0
Flask~=2.0.1
Flask-RESTful~=0.3.9
flake8==3.7.8
pytest==6.2.4
black==21.7b0
networkx~=2.7.1
Jinja2~=3.0.1
gunicorn~=20.1.0
Flask-Cors~=3.0.10
marshmallow==3.14.1
pydantic~=1.9.0
numpydoc~=1.1.0
PyYAML~=5.4.1
requests~=2.27.1
